Combine actions common and index
